+++ jest.envSetup.js
@@ -1,0 +1,8 @@
+global.TERRA_AGGREGATED_LOCALES = ['en', 'es', 'de'];
+global.TERRA_THEME_CONFIG = {
+  scoped: [
+    'clinical-lowlight-theme',
+    'orion-fusion-theme',
+  ],
+  theme: 'orion-fusion-theme',
+};